[hotfix] Fix spark test type and missing test
