+++ paimon-spark/paimon-spark-3.2/src/test/scala/org/apache/paimon/spark/procedure/ProcedureTest.scala
@@ -15,19 +15,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.paimon.spark.sql
+package org.apache.paimon.spark.procedure
 
-import org.apache.spark.sql.catalyst.dsl.expressions._
-import org.apache.spark.sql.catalyst.expressions.{Attribute, GetStructField, NamedExpression, ScalarSubquery}
-import org.apache.spark.sql.catalyst.plans.logical.CTERelationRef
-
-class PaimonOptimzationTest extends PaimonOptimzationTestBase {
-
-  override def extractorExpression(
-      cteIndex: Int,
-      output: Seq[Attribute],
-      fieldIndex: Int): NamedExpression = {
-    GetStructField(ScalarSubquery(CTERelationRef(cteIndex, _resolved = true, output)), fieldIndex)
-      .as("scalarsubquery()")
-  }
-}
+class ProcedureTest extends ProcedureTestBase {}
