+++ paimon-spark/paimon-spark-common/src/test/scala/org/apache/paimon/spark/sql/DeleteFromTableTest.scala
@@ -33,7 +33,7 @@ class DeleteFromTableTest extends PaimonSparkTestBase {
     spark.sql("INSERT INTO T VALUES (1, 'a', '11'), (2, 'b', '22')")
 
     assertThatThrownBy(() => spark.sql("DELETE FROM T WHERE name = 'a'"))
-      .isInstanceOf(classOf[UnsupportedOperationException]);
+      .isInstanceOf(classOf[UnsupportedOperationException])
   }
 
   CoreOptions.MergeEngine.values().foreach {
